Делает домашнее задание 5 модуля
